pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-kgqa-env"
version = "0.1.0"
description = "KGQA Environment for OpenEnv - Knowledge Graph construction, repair, and reasoning"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "fastapi>=0.115.0",
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
baseline = [
    "openai>=1.0.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
]
